requirements.txt
amqp==2.6.0
aniso8601==7.0.0
asgiref==3.2.10
autopep8==1.5.3
Babel==2.8.0
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
dj-email-url==1.0.1
Django==3.1.2
django-appconf==1.0.4
django-cors-headers==3.5.0
django-countries==6.1.2
django-filter==2.4.0
django-graphql-jwt==0.3.0
django-measurement==3.2.3
django-phonenumber-field==4.0.0
django-prices==2.2.0
django-render-block==0.6
django-templated-email==2.3.0
django-versatileimagefield==2.0
draftjs-sanitizer==1.0.0
enmerkar==0.7.1
future==0.18.2
google-i18n-address==2.3.5
graphene==2.1.8
graphene-django==2.11.0
graphene-federation==0.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
kombu==4.6.11
measurement==3.2.0
mpmath==1.1.0
opentracing==2.3.0
phonenumberslite==8.12.6
Pillow==7.1.2
prices==1.0.0
promise==2.3
psycopg2-binary==2.8.5
psycopg2==2.8.5
pycodestyle==2.6.0
PyJWT==1.7.1
python-magic==0.4.18
python-magic-bin==0.4.14
pytimeparse==1.1.8
pytz==2020.1
requests==2.24.0
Rx==1.6.1
sentry-sdk==0.16.0
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
sympy==1.6.1
toml==0.10.1
typing==3.7.4.1
Unidecode==1.1.1
urllib3==1.25.9
vine==1.3.0
